Add PSM support to demo
Add Power Save Mode (PSM) to the demo.
The PSM active timer is set to 16 seconds.  The PSM periodic timer is set to 90 seconds.  The app is set to send sensor data every 90 seconds.

Update the README to reflect the new timings.  Add a table of contents.

Turn on MQTT QOS 1 (ACKS).

Bump the firmware revision in preparation for release.

bug 14384