Disable eDRX
There are currently issues with eDRX in the HL7800 driver that do not allow the HL7800 to stay connected to AWS when it enters low power mode.  The OOB demo is designed to stay connected to AWS in case the server (Bluegrass web app) needs to send messages to the dev board.

bug 16750